[adding_sync] Refactoring api.ws and moving files
